Object structure defining
The project of this is built. Please commit the change with the
following guideline.

1. For main activity, leave it blank for now as later on it will be
used as integration page. If you want to test the part you are in
charge, please create a new activity and test it there at the moment.

2. The object structure are defined, including variables as well as
some methods. If you wish to change the variable that is already
defined, please includes the variable name change in the description
when you commit the change to github. If just adding variable and
method, then no need.

3. Update the change only when your part does not contain any errors.
As it might affect others code when the sync with the online version.